Logout and image router testing

+++ backend/src/controllers/logout.js
@@ -1,14 +1,13 @@
-const jwt = require('jsonwebtoken')
-const bcrypt = require('bcrypt')
-const loginRouter = require('express').Router()
-const User = require('../models/user')
+const logoutRouter = require('express').Router()
 const Session = require('../models/session')
-const middleware = require('../utils/middleware')
+const jwt = require('jsonwebtoken')
 
-loginRouter.delete('/', async (request, response, next) => {
+logoutRouter.delete('/', async (request, response, next) => {
 	try {
 		if (request.token) {
-			await Session.deleteMany({ token: request.token })
+			const decodedToken = jwt.verify(request.token, process.env.SECRET)
+			const session = decodedToken.id ? await Session.findOne({ token: request.token }) : undefined
+			session ? await Session.deleteMany({ userId: session.userId }) : await Session.deleteMany({ token: request.token })
 		}
 		response.status(204).end()
 	} catch (exception) {
@@ -16,4 +15,4 @@ loginRouter.delete('/', async (request, response, next) => {
 	}
 })
 
-module.exports = loginRouter
+module.exports = logoutRouter
